+++ src/main/java/com/skaggsm/jmumblelink/MumbleLinkMemory.java
@@ -1,4 +1,4 @@
-package com.skaggsm.mumblelink;
+package com.skaggsm.jmumblelink;
 
 import com.sun.jna.Pointer;
 import com.sun.jna.Structure;
