+++ version.properties
@@ -1,4 +1,4 @@
 #Version of the produced binaries. This file is intended to be checked-in.
 #It will be automatically bumped by release automation.
-version=0.1.1
+version=0.2.0
 previousVersion=0.1.0
